+++ lib/map_stupidshared.c
@@ -62,8 +62,8 @@ map_refresh(fd, onceonly, base, len, newlen, name, mboxname)
 int fd;
 int onceonly;
 const char **base;
-unsigned long *len;
-unsigned long newlen;
+size_t *len;
+size_t newlen;
 const char *name;
 const char *mboxname;
 {
@@ -108,7 +108,7 @@ const char *mboxname;
 void
 map_free(base, len)
 const char **base;
-unsigned long *len;
+size_t *len;
 {
     if (*len) munmap((char *)*base, *len);
     *base = 0;
